Added a GameOverMenu-UI and fix the POST-request in The HighScoreService

+++ Assets/Scripts/Scriptable Objects/IntVariable.cs
@@ -5,8 +5,8 @@ namespace FG {
 	public class IntVariable : ScriptableObject
 	{
 		/// <summary>
-		/// The float value that is being used
+		/// The int value that is being used
 		/// </summary>
-		public float Value { get; set; }
+		public int Value { get; set; }
 	}
 }
